Simplify functional test a bit more
